requirements.txt
crc16==0.1.1
gpiozero==1.5.0
h5py==2.10.0
https://www.piwheels.org/simple/matplotlib/matplotlib-3.1.1-cp37-cp37m-linux_armv7l.whl
https://www.piwheels.org/simple/numpy/numpy-1.17.0rc1-cp37-cp37m-linux_armv7l.whl
iso8601==0.1.12
logzero==1.5.0
pandas==0.25.3
pyserial==3.4
pytz==2018.9
tendo==0.2.12
xmodem==0.4.5
pyyaml==5.3.1
fiona==1.8.13.post1
cartopy==0.18.0
tzlocal==2.1
wifi==0.3.8

# on rpi, pyqt5 w/ pip crashes, do apt-get
# bluepy mod installed by ddh scripts in lowell_tools repo
# lowell-mat installed by ddh scripts in lowell_tools repo